Move and reorganize code into packages
Made some variables private
