+++ src/main/java/kidsnd274/duke/exceptions/DukeException.java
@@ -1,4 +1,4 @@
-package DukeException;
+package kidsnd274.duke.exceptions;
 
 public class DukeException extends Exception {
     public DukeException(String str) {
