+++ text-ui-test/EXPECTED.TXT
@@ -1,6 +1,6 @@
 Hello from
- ____        _
-|  _ \ _   _| | _____
+ ____        _        
+|  _ \ _   _| | _____ 
 | | | | | | | |/ / _ \
 | |_| | |_| |   <  __/
 |____/ \__,_|_|\_\___|
